Add user collation support
Add support to create and delete [custom
collations](https://www.sqlite.org/c3ref/create_collation.html).

The functionality is similair to user functions, but there's a fixed arity (two
strings) and a fixed return value (type) so the code is quite simple compared to the
implementation of user functions.

Automated test is included to ensure that the functionality actually works as expected.